requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
django-markdownx==4.0.9
djangorestframework==3.16.1
Markdown==3.10
pillow==12.1.0
sqlparse==0.5.5
